+++ yospace/src/main/java/com/bitmovin/player/integration/yospace/config/TruexConfig.kt
@@ -2,7 +2,7 @@ package com.bitmovin.player.integration.yospace.config
 
 import android.view.ViewGroup
 
-data class TruexConfiguration(
+data class TruexConfig(
     val viewGroup: ViewGroup,
     val userId: String? = null,
     val vastConfigUrl: String? = null