+++ yospacesample/src/main/java/com/bitmovin/player/integration/yospacesample/MainActivity.kt
@@ -2,17 +2,20 @@ package com.bitmovin.player.integration.yospacesample
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import com.bitmovin.player.api.event.listener.OnSourceLoadedListener
-import com.bitmovin.player.api.event.listener.OnSourceUnloadedListener
-import com.bitmovin.player.config.PlayerConfiguration
-import com.bitmovin.player.config.drm.WidevineConfiguration
-import com.bitmovin.player.config.media.HLSSource
-import com.bitmovin.player.config.media.SourceConfiguration
-import com.bitmovin.player.config.media.SourceItem
+import com.bitmovin.player.api.PlayerConfig
+import com.bitmovin.player.api.PlaybackConfig
+import com.bitmovin.player.api.Player
+import com.bitmovin.player.api.TweaksConfig
+import com.bitmovin.player.api.drm.WidevineConfig
+import com.bitmovin.player.api.event.SourceEvent
+import com.bitmovin.player.api.event.on
+import com.bitmovin.player.api.source.SourceConfig
+import com.bitmovin.player.api.source.SourceType
+import com.bitmovin.player.integration.yospace.BitLog
 import com.bitmovin.player.integration.yospace.BitmovinYospacePlayer
 import com.bitmovin.player.integration.yospace.YospaceAssetType
-import com.bitmovin.player.integration.yospace.config.YospaceConfiguration
-import com.bitmovin.player.integration.yospace.config.YospaceSourceConfiguration
+import com.bitmovin.player.integration.yospace.config.YospaceConfig
+import com.bitmovin.player.integration.yospace.config.YospaceSourceConfig
 import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
@@ -24,18 +27,18 @@ class MainActivity : AppCompatActivity() {
             Stream(
                 "Yospace Live",
                 "https://csm-e-sdk-validation.bln1.yospace.com/csm/extlive/yospace02,hlssample42.m3u8?yo.br=true&yo.av=4",
-                yospaceSourceConfig = YospaceSourceConfiguration(YospaceAssetType.LINEAR)
+                yospaceSourceConfig = YospaceSourceConfig(YospaceAssetType.LINEAR)
             ),
             Stream(
                 "Yospace Companion Ads",
                 "https://csm-e-sdk-validation.bln1.yospace.com/csm/extlive/yospace02,hlssample42.m3u8?yo.br=true&yo.lp=true&yo.av=4",
                 "https://widevine-stage.license.istreamplanet.com/widevine/api/license/de4c1d30-ac22-4669-8824-19ba9a1dc128",
-                YospaceSourceConfiguration(YospaceAssetType.LINEAR)
+                YospaceSourceConfig(YospaceAssetType.LINEAR)
             ),
             Stream(
                 "Yospace VOD",
-                "https://csm-e-sdk-validation.bln1.yospace.com/csm/access/156611618/c2FtcGxlL21hc3Rlci5tM3U4?yo.av=4",
-                yospaceSourceConfig = YospaceSourceConfiguration(YospaceAssetType.VOD)
+                "https://csm-e-sdk-validation.bln1.yospace.com/csm/access/207411697/c2FtcGxlL21hc3Rlci5tM3U4?yo.av=3",
+                yospaceSourceConfig = YospaceSourceConfig(YospaceAssetType.VOD)
             )
         )
     }
@@ -67,48 +70,43 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun setupPlayer() {
-        val playerConfig = PlayerConfiguration().apply {
-            playbackConfiguration?.isAutoplayEnabled = true
-            tweaksConfiguration?.useFiletypeExtractorFallbackForHls = true
-        }
+        val playerConfig = PlayerConfig(
+            playbackConfig = PlaybackConfig(isAutoplayEnabled = true),
+            tweaksConfig = TweaksConfig(useFiletypeExtractorFallbackForHls = true)
+        )
 
-        player = BitmovinYospacePlayer(this, playerConfig, YospaceConfiguration()).apply {
-            addEventListener(OnSourceLoadedListener {
-                loadUnloadButton.text = getString(R.string.unload)
-            })
-            addEventListener(OnSourceUnloadedListener {
-                loadUnloadButton.text = getString(R.string.load)
-            })
-            playerView.player = this
+        player = BitmovinYospacePlayer(this, playerConfig, yospaceConfig = YospaceConfig())
+        player.on<SourceEvent.Load> {
+            BitLog.d("Change button")
+            loadUnloadButton.text = getString(R.string.unload)
         }
+        player.on<SourceEvent.Unloaded> {
+            loadUnloadButton.text = getString(R.string.load)
+        }
+        playerView.player = player
+        BitLog.d("Setup player")
     }
 
     private fun setupSpinner() {
-        streamSpinner.adapter = StreamSpinnerAdapter(this, streams, android.R.layout.simple_spinner_item)
+        streamSpinner.adapter =
+            StreamSpinnerAdapter(this, streams, android.R.layout.simple_spinner_item)
     }
 
     private fun addUIListeners() {
         loadUnloadButton.setOnClickListener {
             if (player.isPlaying) {
+                BitLog.d("unload stream")
                 player.unload()
             } else {
+                BitLog.d("Button clicked, load stream")
                 loadStream(streams[streamSpinner.selectedItemPosition])
             }
         }
     }
 
     private fun loadStream(stream: Stream) {
-        val sourceItem = SourceItem(
-            HLSSource(stream.contentUrl)
-        )
-
-        stream.drmUrl?.let {
-            sourceItem.addDRMConfiguration(WidevineConfiguration(it))
-        }
-
-        val sourceConfig = SourceConfiguration().apply {
-            addSourceItem(sourceItem)
-        }
+        val sourceConfig = SourceConfig(stream.contentUrl, SourceType.Hls)
+        sourceConfig.drmConfig = WidevineConfig(stream.drmUrl)
 
         player.load(sourceConfig, stream.yospaceSourceConfig, stream.truexConfig)
     }